+++ bin/env_reader.dart
@@ -2,7 +2,7 @@
 import 'dart:io';
 
 import 'package:args/args.dart';
-import 'package:encryptor/encryptor.dart';
+import 'package:env_reader/env_reader_core.dart' show EnvEncryption;
 
 part 'src/file.dart';
 part 'src/pubspec.dart';
@@ -23,6 +23,10 @@ void main(List<String> arguments) async {
         defaultsTo: false, negatable: false, help: 'Make the model null safety')
     ..addFlag('obfuscate',
         defaultsTo: true, help: 'Obfuscating generated values of model')
+    ..addOption('sdk',
+        help: 'Choose between generating model for flutter or dart project',
+        allowed: ['dart', 'flutter'],
+        defaultsTo: 'flutter')
     ..addFlag('pubspec',
         defaultsTo: true, help: 'Inserting asset path to pubspec.yaml')
     ..addFlag('gitignore',