+++ example/lib/src/env_model.dart
@@ -1,5 +1,5 @@
 // Env Reader Auto-Generated Model File
-// Created at 2023-10-11 15:28:35.710424
+// Created at 2023-10-16 16:30:38.306175
 // 🍔 [Buy me a coffee](https://www.buymeacoffee.com/nialixus) 🚀
 import 'package:env_reader/env_reader.dart';